Remove raven middleware from client and server applications
This disables mobile-web error reporting to Sentry. The goal
was to make the smallest change necessary to make sure errors
weren't being reported anymore as mobile-web is currently not
being actively worked on and existing errors, which make up
a disproportional amount of error-reporting traffic, are not
being addressed. This should allow for quick re-enabling in the
future if desired.