Add best sort for logged in users & fix sorts on search page
Previously, sorting on the search page was totally broken, and would navigate the user back to the subreddit page instead of just changing the sort. This commit fixes that bug, and also sets 'best' as the default sort option for the home feed (logged in users only).

This commit also does some refactoring and adds a concept of PAGE_NAMES so you can check the type of page you are on.